requirements.txt
appdirs==1.4.4
aspy.yaml==1.3.0
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bpyutils==0.2.0
business-rules==1.0.1
cached-property==1.5.2
ccad==2.0
certifi==2021.5.30
cfgv==3.3.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
decorator==5.1.0
distlib==0.3.3
euclid3==0.1
filelock==3.3.0
freetype-py==2.2.0
future==0.18.2
graphviz==0.17
hsluv==5.0.2
identify==2.3.0
importlib-metadata==4.8.1
iniconfig==1.1.1
jedi==0.18.0
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
more-itertools==8.10.0
mypy==0.910
mypy-extensions==0.4.3
networkx==2.6.3
nodeenv==1.6.0
numpy==1.21.2
packaging==21.0
parso==0.8.2
Pillow==8.3.2
pipupgrade==1.10.1
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.15.0
prettytable==0.7.2
py==1.10.0
pyglet==1.5.21
pyparsing==2.4.7
pypng==0.0.19
PyQt5==5.15.4
PyQt5-sip==12.9.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.10.8
scipy==1.7.1
sip==6.2.0
six==1.16.0
solidpython==1.0.2
toml==0.10.2
trimesh==3.9.32
typed-ast==1.4.3
typing-extensions==3.10.0.2
virtualenv==20.8.1
vispy==0.6.6
wcwidth==0.2.5
zipp==3.6.0

